==> notebooks/06_version2/01_simulation_datasets/03b_GTDBr202/01a_n9k_train_CAMI-err.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Goal" data-toc-modified-id="Goal-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Goal</a></span></li><li><span><a href="#Var" data-toc-modified-id="Var-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Var</a></span></li><li><span><a href="#Init" data-toc-modified-id="Init-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Init</a></span></li><li><span><a href="#ResMiCo-sm" data-toc-modified-id="ResMiCo-sm-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>ResMiCo-sm</a></span><ul class="toc-item"><li><span><a href="#Run1" data-toc-modified-id="Run1-4.1"><span class="toc-item-num">4.1&nbsp;&nbsp;</span>Run1</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-4.1.1"><span class="toc-item-num">4.1.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-4.1.2"><span class="toc-item-num">4.1.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li><li><span><a href="#Run2" data-toc-modified-id="Run2-4.2"><span class="toc-item-num">4.2&nbsp;&nbsp;</span>Run2</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-4.2.1"><span class="toc-item-num">4.2.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-4.2.2"><span class="toc-item-num">4.2.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li><li><span><a href="#Run3" data-toc-modified-id="Run3-4.3"><span class="toc-item-num">4.3&nbsp;&nbsp;</span>Run3</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-4.3.1"><span class="toc-item-num">4.3.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-4.3.2"><span class="toc-item-num">4.3.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li><li><span><a href="#Creating-the-features-file-table" data-toc-modified-id="Creating-the-features-file-table-4.4"><span class="toc-item-num">4.4&nbsp;&nbsp;</span>Creating the features file table</a></span></li></ul></li><li><span><a href="#Sandbox" data-toc-modified-id="Sandbox-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Sandbox</a></span></li><li><span><a href="#sessionInfo" data-toc-modified-id="sessionInfo-6"><span class="toc-item-num">6&nbsp;&nbsp;</span>sessionInfo</a></span></li></ul></div>

# Goal

* GTDBr202 n9k train dataset with CAMI-error rates for simulated reads
  * Using error rates obtained from the [CAMISIM repo](https://github.com/CAMI-challenge/CAMISIM)

# Var

work_dir = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_CAMI-err/'

ref_gen_file = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/ref_genomes/GTDBr202/metadata_n9k-train.tsv'

# params
threads = 12

# Init

library(dplyr)
library(tidyr)
library(ggplot2)
library(data.table)
library(tidytable)
library(LeyLabRMisc)

library(PRROC)

clustermq_setup(scheduler = c("multicore"))
df.dims()
setDTthreads(threads)
make_dir(work_dir)

# ResMiCo-sm

## Run1

### Config

F = file.path(work_dir, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/bin/ResMiCo/ResMiCo-SM
$ screen -L -S rmc-sm-GTDBr202-train ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_CAMI-err/config.yaml 60 --notemp -F
```

## Run2

* More simulation parameters

### Config

base_dir = file.path(dirname(work_dir), 'GTDBr202_n9k_train_CAMI-err_d20m')
F = file.path(base_dir, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/bin/ResMiCo/ResMiCo-SM
$ screen -L -S rmc-sm-GTDBr202-train ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_CAMI-err_d20m/config.yaml 60 --notemp -F
```

## Run3

* Extra simulation replicates

### Config

base_dir = file.path(dirname(work_dir), 'GTDBr202_n9k_train_CAMI-err_r2r3')
F = file.path(base_dir, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/bin/new_compile/ResMiCo/ResMiCo-SM
$ screen -L -S rmc-sm-cami-err ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_CAMI-err_r2r3/config.yaml 40
```

## Creating the features file table

* snakemake failed to create 6 of the 480 feature table files
* I now need to "manually" create the table that lists all of the feature table files

D = file.path(base_dir, 'features')
files = list_files(D, 'features.tsv.gz')
files %>% length
files %>% head(n=3)

get_value = function(x, i){
    x = unlist(strsplit(x, '/'))
    return(x[i])
}

j = 10
df = data.frame(
    richness = sapply(files, get_value, i=j),
    abundance_distribution = sapply(files, get_value, i=j+1),
    rep = sapply(files, get_value, i=j+2),
    read_length = sapply(files, get_value, i=j+3),
    read_depth = sapply(files, get_value, i=j+4),
    assembler = sapply(files, get_value, i=j+5),
    feature_file = files,
    MGSIM_reads_params = NA
)
rownames(df) = 1:nrow(df)
df

F = file.path(base_dir, 'features', 'feature_files.tsv')
write_table(df, F)

# Sandbox

#base_dir = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test2/'

D = file.path(base_dir, 'features')
files = list_files(D, 'features.tsv.gz')
files %>% length
files %>% head(n=3)

get_value = function(x, i){
    x = unlist(strsplit(x, '/'))
    return(x[i])
}

j = 10
df = data.frame(
    richness = sapply(files, get_value, i=j),
    abundance_distribution = sapply(files, get_value, i=j+1),
    rep = sapply(files, get_value, i=j+2),
    read_length = sapply(files, get_value, i=j+3),
    read_depth = sapply(files, get_value, i=j+4),
    assembler = sapply(files, get_value, i=j+5),
    feature_file = files,
    MGSIM_reads_params = NA
)
rownames(df) = 1:nrow(df)
df

F = file.path(base_dir, 'features', 'feature_files.tsv')
write_table(df, F)

# sessionInfo

sessionInfo()
