==> notebooks/06_version2/01_simulation_datasets/03b_GTDBr202/01b_n9k_train_large-insert.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Goal" data-toc-modified-id="Goal-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Goal</a></span></li><li><span><a href="#Var" data-toc-modified-id="Var-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Var</a></span></li><li><span><a href="#Init" data-toc-modified-id="Init-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Init</a></span></li><li><span><a href="#ResMiCo-sm" data-toc-modified-id="ResMiCo-sm-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>ResMiCo-sm</a></span><ul class="toc-item"><li><span><a href="#Run1:-CAMI-error" data-toc-modified-id="Run1:-CAMI-error-4.1"><span class="toc-item-num">4.1&nbsp;&nbsp;</span>Run1: CAMI-error</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-4.1.1"><span class="toc-item-num">4.1.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-4.1.2"><span class="toc-item-num">4.1.2&nbsp;&nbsp;</span>Run</a></span></li><li><span><a href="#Summary" data-toc-modified-id="Summary-4.1.3"><span class="toc-item-num">4.1.3&nbsp;&nbsp;</span>Summary</a></span><ul class="toc-item"><li><span><a href="#Feature-file-table" data-toc-modified-id="Feature-file-table-4.1.3.1"><span class="toc-item-num">4.1.3.1&nbsp;&nbsp;</span>Feature file table</a></span></li><li><span><a href="#Missing?" data-toc-modified-id="Missing?-4.1.3.2"><span class="toc-item-num">4.1.3.2&nbsp;&nbsp;</span>Missing?</a></span></li></ul></li></ul></li><li><span><a href="#Run2:-standard-error" data-toc-modified-id="Run2:-standard-error-4.2"><span class="toc-item-num">4.2&nbsp;&nbsp;</span>Run2: standard-error</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-4.2.1"><span class="toc-item-num">4.2.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-4.2.2"><span class="toc-item-num">4.2.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li></ul></li><li><span><a href="#ResMiCo-sm:-reps-(2,3)" data-toc-modified-id="ResMiCo-sm:-reps-(2,3)-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>ResMiCo-sm: reps (2,3)</a></span><ul class="toc-item"><li><span><a href="#CAMI-error" data-toc-modified-id="CAMI-error-5.1"><span class="toc-item-num">5.1&nbsp;&nbsp;</span>CAMI-error</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-5.1.1"><span class="toc-item-num">5.1.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-5.1.2"><span class="toc-item-num">5.1.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li><li><span><a href="#Standard-error" data-toc-modified-id="Standard-error-5.2"><span class="toc-item-num">5.2&nbsp;&nbsp;</span>Standard error</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-5.2.1"><span class="toc-item-num">5.2.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-5.2.2"><span class="toc-item-num">5.2.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li></ul></li><li><span><a href="#--WAITING--" data-toc-modified-id="--WAITING---6"><span class="toc-item-num">6&nbsp;&nbsp;</span>--WAITING--</a></span></li><li><span><a href="#sessionInfo" data-toc-modified-id="sessionInfo-7"><span class="toc-item-num">7&nbsp;&nbsp;</span>sessionInfo</a></span></li></ul></div>

# Goal

* GTDBr202 n9k train dataset with CAMI-error rates for simulated reads
  * Using error rates obtained from the [CAMISIM repo](https://github.com/CAMI-challenge/CAMISIM)
* Using a larger insert size distribution used in the original simulations

# Var

work_dir = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_lrg-ins_CAMI-err/'

ref_gen_file = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/ref_genomes/GTDBr202/metadata_n9k-train.tsv'

# params
threads = 12

# Init

library(dplyr)
library(tidyr)
library(ggplot2)
library(data.table)
library(tidytable)
library(LeyLabRMisc)

library(PRROC)

clustermq_setup(scheduler = c("multicore"))
df.dims()
setDTthreads(threads)
make_dir(work_dir)

# ResMiCo-sm

## Run1: CAMI-error

* cami error profiles

### Config

F = file.path(work_dir, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/bin/old_compile/ResMiCo/ResMiCo-SM
$ screen -L -S rmc-sm-lrgIns ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_lrg-ins_CAMI-err/config.yaml 50 -F
```

### Summary

* number of total parameter combinations: 240

#### Feature file table

feat_files = list_files(file.path(work_dir, 'features'), 'features.tsv.gz')
feat_files %>% length
feat_files %>% head(n=3)

parse_path = function(F){
    p = F %>% gsub('/+', '/', .) %>% strsplit('/') %>% unlist
    d = data.frame(
        'richness' = p[10],
        'abundance_distribution' = p[11], 
        'rep' = p[12], 
        'read_length' = p[13], 
        'read_depth' = p[14], 
        'assembler' = p[15],
        'feature_file' = F, 
        'MGSIM_reads_params' = 'NA'
    )
    return(d)
}
feat_files = feat_files %>%
    lapply(parse_path) %>%
    rbindlist(use.names=TRUE)
feat_files

F = file.path(work_dir, 'features', 'feature_files.tsv')
write_table(feat_files, F)

#### Missing?

richness = c(0.005555, 0.111111, 0.333333, 0.555555)
abund_dist = c("mean-10-sigma-0.5", "mean-10-sigma-1", "mean-10-sigma-2")
read_length = c(100, 150)
read_depth = c(500000, 2000000, 8000000, 12000000, 20000000)
assembler = c('megahit', 'metaspades')
all_params = expand.grid(richness, abund_dist, read_length, read_depth, assembler)
colnames(all_params) = c('richness', 'abundance_distribution', 'read_length', 'read_depth', 'assembler')
all_params

# filtering
all_params_f = all_params %>%
    anti_join.(feat_files %>%
                   mutate.(richness = as.Num(richness),
                           read_length = as.Num(read_length),
                           read_depth = as.Num(read_depth)),
               c('richness', 'abundance_distribution', 'read_length', 'read_depth', 'assembler'))
all_params_f

df.dims(20)
all_params_f %>% arrange.(richness, abundance_distribution, read_length, read_depth, assembler)
df.dims()

F = file.path(work_dir, 'missing_params.tsv')
all_params_f %>% 
    rename.('abund_dist' = abundance_distribution) %>%
    write_table(F)

## Run2: standard-error

* "standard" error profile

D = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_lrg-ins/'

### Config

F = file.path(D, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico- sm/bin/old_compile/ResMiCo/ResMiCo-SM
$ screen -L -S rmc-sm-lrgIns ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_lrg-ins/config.yaml 50 -F
```

# ResMiCo-sm: reps (2,3)

D = file.path(dirname(work_dir), 'GTDBr202_n9k_train_lrg-ins_CAMI-err_r23')
make_dir(D)

## CAMI-error

### Config

F = file.path(D, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/bin/old_compile/ResMiCo/ResMiCo-SM
$ screen -L -S rmc-sm-lrgIns ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_lrg-ins_CAMI-err_r23/config.yaml 40 -F
```

## Standard error

D = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_lrg-ins_r23/'

### Config

F = file.path(D, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) rick: /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/bin/old_compile/ResMiCo/ResMiCo-SM
$ screen -L -S rmc-sm-lrgIns ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_train_lrg-ins_r23/config.yaml 30 -F
```

# --WAITING--

# sessionInfo

sessionInfo()
