==> notebooks/06_version2/01_simulation_datasets/03b_GTDBr202/03_n9k_novel-family_test.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Goal" data-toc-modified-id="Goal-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Goal</a></span></li><li><span><a href="#Var" data-toc-modified-id="Var-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Var</a></span></li><li><span><a href="#Init" data-toc-modified-id="Init-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Init</a></span></li><li><span><a href="#Check" data-toc-modified-id="Check-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Check</a></span><ul class="toc-item"><li><ul class="toc-item"><li><ul class="toc-item"><li><span><a href="#Checking-overlap-with-train" data-toc-modified-id="Checking-overlap-with-train-4.0.0.1"><span class="toc-item-num">4.0.0.1&nbsp;&nbsp;</span>Checking overlap with train</a></span></li><li><span><a href="#Writing-table" data-toc-modified-id="Writing-table-4.0.0.2"><span class="toc-item-num">4.0.0.2&nbsp;&nbsp;</span>Writing table</a></span></li></ul></li></ul></li></ul></li><li><span><a href="#ResMiCo-Sm" data-toc-modified-id="ResMiCo-Sm-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>ResMiCo-Sm</a></span><ul class="toc-item"><li><span><a href="#Run1" data-toc-modified-id="Run1-5.1"><span class="toc-item-num">5.1&nbsp;&nbsp;</span>Run1</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-5.1.1"><span class="toc-item-num">5.1.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-5.1.2"><span class="toc-item-num">5.1.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li><li><span><a href="#Run2" data-toc-modified-id="Run2-5.2"><span class="toc-item-num">5.2&nbsp;&nbsp;</span>Run2</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-5.2.1"><span class="toc-item-num">5.2.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-5.2.2"><span class="toc-item-num">5.2.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li></ul></li><li><span><a href="#sessionInfo" data-toc-modified-id="sessionInfo-6"><span class="toc-item-num">6&nbsp;&nbsp;</span>sessionInfo</a></span></li></ul></div>

# Goal

* ResMiCo-SM run on `GTDBr202_n9k_test` dataset

# Var

work_dir = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test/'

# reference genomes
## test
ref_gen_file = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/ref_genomes/GTDBr202/metadata_n9k-novel-family.tsv'
## train
ref_gen_train_file = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/ref_genomes/GTDBr202/metadata_n9k-train.tsv'

# params
threads = 12

# Init

library(dplyr)
library(tidyr)
library(ggplot2)
library(data.table)
library(tidytable)
library(LeyLabRMisc)

library(PRROC)

clustermq_setup(scheduler = c("multicore"))
df.dims()
setDTthreads(threads)
make_dir(work_dir)

# Check

# reference genomes
ref_gen = Fread(ref_gen_file)
ref_gen

# number of genomes
ref_gen$accession %>% unique_n('genomes')
ref_gen$ncbi_genbank_assembly_accession %>% unique_n('NCBI accessions')

# number of species
ref_gen$Species %>% unique_n('species')

#### Checking overlap with train

ref_gen_train = Fread(ref_gen_train_file)
ref_gen_train

# overlapping genomes?
overlap(ref_gen$accession, ref_gen_train$accession)

# overlapping taxonomies?
cat('---Order--\n')
overlap(ref_gen$Order, ref_gen_train$Order)
cat('---Family--\n')
overlap(ref_gen$Family, ref_gen_train$Family)
cat('---Genus--\n')
overlap(ref_gen$Genus, ref_gen_train$Genus)

#### Writing table

F = file.path(work_dir, 'genomes.tsv')
ref_gen %>%
    rename.('Taxon' = accession,
            'Fasta' = fasta_file_path) %>%
    write_table(F)

# ResMiCo-Sm

## Run1

### Config

F = file.path(work_dir, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/bin/ResMiCo/ResMiCo-SM
$ screen -L -S rmc-sm-test ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test/config.yaml 40 -F --notemp
```

## Run2

* Expanding simulation parameters

work_dir2 = paste0(gsub('/$', '', work_dir), '2')
work_dir2

### Config

F = file.path(work_dir2, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/bin/ResMiCo/ResMiCo-SM
$ screen -L -S rmc-sm-test ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test2/config.yaml 40 -F --notemp
```

# sessionInfo

sessionInfo()
