==> notebooks/06_version2/01_simulation_datasets/03b_GTDBr202/04_n9k_diff-kmer.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Goal" data-toc-modified-id="Goal-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Goal</a></span></li><li><span><a href="#Var" data-toc-modified-id="Var-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Var</a></span></li><li><span><a href="#Init" data-toc-modified-id="Init-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Init</a></span></li><li><span><a href="#ResMiCo-sm" data-toc-modified-id="ResMiCo-sm-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>ResMiCo-sm</a></span><ul class="toc-item"><li><span><a href="#kmer21-31-41" data-toc-modified-id="kmer21-31-41-4.1"><span class="toc-item-num">4.1&nbsp;&nbsp;</span>kmer21-31-41</a></span><ul class="toc-item"><li><span><a href="#Run" data-toc-modified-id="Run-4.1.1"><span class="toc-item-num">4.1.1&nbsp;&nbsp;</span>Run</a></span></li></ul></li><li><span><a href="#kmer51-61-71" data-toc-modified-id="kmer51-61-71-4.2"><span class="toc-item-num">4.2&nbsp;&nbsp;</span>kmer51-61-71</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-4.2.1"><span class="toc-item-num">4.2.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-4.2.2"><span class="toc-item-num">4.2.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li><li><span><a href="#kmer91-101-111" data-toc-modified-id="kmer91-101-111-4.3"><span class="toc-item-num">4.3&nbsp;&nbsp;</span>kmer91-101-111</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-4.3.1"><span class="toc-item-num">4.3.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-4.3.2"><span class="toc-item-num">4.3.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li><li><span><a href="#kmer21-41-61" data-toc-modified-id="kmer21-41-61-4.4"><span class="toc-item-num">4.4&nbsp;&nbsp;</span>kmer21-41-61</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-4.4.1"><span class="toc-item-num">4.4.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-4.4.2"><span class="toc-item-num">4.4.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li><li><span><a href="#kmer51-71-91" data-toc-modified-id="kmer51-71-91-4.5"><span class="toc-item-num">4.5&nbsp;&nbsp;</span>kmer51-71-91</a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-4.5.1"><span class="toc-item-num">4.5.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-4.5.2"><span class="toc-item-num">4.5.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li></ul></li><li><span><a href="#sessionInfo" data-toc-modified-id="sessionInfo-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>sessionInfo</a></span></li></ul></div>

# Goal

* ResMiCo-SM runs to test the sensitivity of ResMiCo to varying assembly kmer lengths
  * Using the same communities for all kmer-combinations
* If ResMiCo is no sensitive to kmer lengths, then the model can be used to optimize assemblers

# Var

work_dir = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test_kmer/'

# reference genomes
## test
ref_gen_file = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/ref_genomes/GTDBr202/metadata_n9k-novel-family.tsv'
## train
ref_gen_train_file = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/ref_genomes/GTDBr202/metadata_n9k-train.tsv'

# params
threads = 12

# Init

library(dplyr)
library(tidyr)
library(ggplot2)
library(data.table)
library(tidytable)
library(LeyLabRMisc)

library(PRROC)

clustermq_setup(scheduler = c("multicore"))
df.dims()
setDTthreads(threads)
make_dir(work_dir)

# ResMiCo-sm

## kmer21-31-41

base_dir = file.path(work_dir, 'kmer21-31-41')

F = file.path(base_dir, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/DeepMAsED/version2/deepmased/ResMiCo-SM
screen -L -S rmc-sm-kmer ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test_kmer/kmer21-31-41/config.yaml 30 -F
```

## kmer51-61-71

### Config

base_dir = file.path(work_dir, 'kmer51-61-71')

F = file.path(base_dir, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/DeepMAsED/version2/deepmased/ResMiCo-SM
screen -L -S rmc-sm-kmer ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test_kmer/kmer51-61-71/config.yaml 30 -F
```

## kmer91-101-111

base_dir = file.path(work_dir, 'kmer91-101-111')

### Config

F = file.path(base_dir, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/DeepMAsED/version2/deepmased/ResMiCo-SM
screen -L -S rmc-sm-kmer ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test_kmer/kmer91-101-111/config.yaml 30 -F
```

## kmer21-41-61

base_dir = file.path(work_dir, 'kmer21-41-61')

### Config

F = file.path(base_dir, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/DeepMAsED/version2/deepmased/ResMiCo-SM
screen -L -S rmc-sm-kmer ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test_kmer/kmer21-41-61/config.yaml 30 -F
```

## kmer51-71-91

base_dir = file.path(work_dir, 'kmer51-71-91')

### Config

F = file.path(base_dir, 'config.yaml')
cat_file(F)

### Run

```
(snakemake) @ rick:/ebio/abt3_projects/DeepMAsED/version2/deepmased/ResMiCo-SM
screen -L -S rmc-sm-kmer ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test_kmer/kmer51-71-91/config.yaml 30 -F
```

# sessionInfo

sessionInfo()
