==> notebooks/06_version2/01_simulation_datasets/03b_GTDBr202/07_n9k_novel-family_test_insertSize.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Goal" data-toc-modified-id="Goal-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Goal</a></span></li><li><span><a href="#Var" data-toc-modified-id="Var-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Var</a></span></li><li><span><a href="#Init" data-toc-modified-id="Init-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Init</a></span></li><li><span><a href="#ResMiCo-SM" data-toc-modified-id="ResMiCo-SM-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>ResMiCo-SM</a></span><ul class="toc-item"><li><span><a href="#--art-mflen-270" data-toc-modified-id="--art-mflen-270-4.1"><span class="toc-item-num">4.1&nbsp;&nbsp;</span><code>--art-mflen 270</code></a></span><ul class="toc-item"><li><span><a href="#Config" data-toc-modified-id="Config-4.1.1"><span class="toc-item-num">4.1.1&nbsp;&nbsp;</span>Config</a></span></li><li><span><a href="#Run" data-toc-modified-id="Run-4.1.2"><span class="toc-item-num">4.1.2&nbsp;&nbsp;</span>Run</a></span></li></ul></li></ul></li></ul></div>

# Goal

* n9k_novel-family dataset simulations, but using varying insert size distributions
* testing how insert size distribution affects the model accuracy

# Var

work_dir = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/GTDBr202_n9k_novel-family_test_insertSize/'

# Init

library(dplyr)
library(tidyr)
library(ggplot2)
library(data.table)
library(tidytable)
library(LeyLabRMisc)
Init()

set.seed(2323)

# ResMiCo-SM

## `--art-mflen 270`

D = file.path(work_dir, 'art-mflen-270')
make_dir(D)

### Config

F = file.path(D, 'config.yaml')
cat_file(F)

### Run

```

```
