==> notebooks/06_version2/01_simulation_datasets/03b_GTDBr202/08a_n9k-train_dataset_check.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Var" data-toc-modified-id="Var-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Var</a></span></li><li><span><a href="#Init" data-toc-modified-id="Init-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Init</a></span></li><li><span><a href="#Features" data-toc-modified-id="Features-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Features</a></span></li><li><span><a href="#Params" data-toc-modified-id="Params-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Params</a></span><ul class="toc-item"><li><ul class="toc-item"><li><span><a href="#1st-insert-size" data-toc-modified-id="1st-insert-size-4.0.1"><span class="toc-item-num">4.0.1&nbsp;&nbsp;</span>1st insert size</a></span></li><li><span><a href="#2nd-insert-size" data-toc-modified-id="2nd-insert-size-4.0.2"><span class="toc-item-num">4.0.2&nbsp;&nbsp;</span>2nd insert size</a></span></li><li><span><a href="#Merged" data-toc-modified-id="Merged-4.0.3"><span class="toc-item-num">4.0.3&nbsp;&nbsp;</span>Merged</a></span></li><li><span><a href="#Missing" data-toc-modified-id="Missing-4.0.4"><span class="toc-item-num">4.0.4&nbsp;&nbsp;</span>Missing</a></span></li></ul></li></ul></li><li><span><a href="#ResMiCo-SM" data-toc-modified-id="ResMiCo-SM-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>ResMiCo-SM</a></span><ul class="toc-item"><li><span><a href="#Feature-tables" data-toc-modified-id="Feature-tables-5.1"><span class="toc-item-num">5.1&nbsp;&nbsp;</span>Feature tables</a></span><ul class="toc-item"><li><span><a href="#mean270-stdev50" data-toc-modified-id="mean270-stdev50-5.1.1"><span class="toc-item-num">5.1.1&nbsp;&nbsp;</span>mean270-stdev50</a></span></li><li><span><a href="#mean350-stdev75" data-toc-modified-id="mean350-stdev75-5.1.2"><span class="toc-item-num">5.1.2&nbsp;&nbsp;</span>mean350-stdev75</a></span></li><li><span><a href="#Overlap" data-toc-modified-id="Overlap-5.1.3"><span class="toc-item-num">5.1.3&nbsp;&nbsp;</span>Overlap</a></span></li><li><span><a href="#Copying-samples" data-toc-modified-id="Copying-samples-5.1.4"><span class="toc-item-num">5.1.4&nbsp;&nbsp;</span>Copying samples</a></span></li><li><span><a href="#Creating-new-feature-table" data-toc-modified-id="Creating-new-feature-table-5.1.5"><span class="toc-item-num">5.1.5&nbsp;&nbsp;</span>Creating new feature table</a></span></li></ul></li></ul></li><li><span><a href="#--OLD--" data-toc-modified-id="--OLD---6"><span class="toc-item-num">6&nbsp;&nbsp;</span>--OLD--</a></span><ul class="toc-item"><li><ul class="toc-item"><li><span><a href="#TODO-samples" data-toc-modified-id="TODO-samples-6.0.1"><span class="toc-item-num">6.0.1&nbsp;&nbsp;</span>TODO samples</a></span></li></ul></li></ul></li><li><span><a href="#sessionInfo" data-toc-modified-id="sessionInfo-7"><span class="toc-item-num">7&nbsp;&nbsp;</span>sessionInfo</a></span></li></ul></div>

# Var

work_dir = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/ftp/n9k-train/'

# params
threads = 4

# Init

library(dplyr)
library(tidyr)
library(ggplot2)
library(data.table)
library(tidytable)
library(LeyLabRMisc)
Init(threads=threads)

# Features

F = file.path(work_dir, 'feature_files.tsv')
train = Fread(F)
train

train$error_profile %>% table
train$insert_size %>% table
train$rep %>% table

# Params

# all pairwise comparisons
params = list(
    rep = c(1,2,3),
    richness = c(0.005555, 0.111111, 0.333333, 0.555555),
    read_length = c(100, 150),
    read_depth = c(500000, 2000000, 8000000, 12000000, 20000000),
    abundance_distribution = c('mean-10-sigma-0.5', 'mean-10-sigma-1', 'mean-10-sigma-2'),
    assembler = c('metaspades', 'megahit'),
    error_profile = c('normal_err', 'cami_err'),
    insert_size = c('mean270-stdev50', 'mean350-stdev75')
)
do.call(expand.grid, params)

### 1st insert size

# all pairwise comparisons
params = list(
    rep = c(1,2,3),
    richness = c(0.005555, 0.111111, 0.333333, 0.555555),
    read_length = c(100, 150),
    read_depth = c(500000, 2000000, 8000000, 12000000, 20000000),
    abundance_distribution = c('mean-10-sigma-0.5', 'mean-10-sigma-1', 'mean-10-sigma-2'),
    assembler = c('metaspades', 'megahit'),
    error_profile = c('normal_err', 'cami_err'),
    insert_size = c('mean270-stdev50')
)
train1 = do.call(expand.grid, params)
train1

### 2nd insert size

# all pairwise comparisons
params = list(
    rep = c(1,2,3),
    richness = c(0.005555, 0.111111, 0.333333, 0.555555),
    read_length = c(100, 150),
    read_depth = c(500000, 2000000, 8000000, 12000000, 20000000),
    abundance_distribution = c('mean-10-sigma-0.5', 'mean-10-sigma-1', 'mean-10-sigma-2'),
    assembler = c('metaspades', 'megahit'),
    error_profile = c('cami_err'),
    insert_size = c('mean350-stdev75')
)
train2a = do.call(expand.grid, params)
train2a

# all pairwise comparisons
params = list(
    rep = c(1),
    richness = c(0.005555, 0.111111, 0.333333, 0.555555),
    read_length = c(100, 150),
    read_depth = c(500000, 2000000, 8000000, 12000000, 20000000),
    abundance_distribution = c('mean-10-sigma-0.5', 'mean-10-sigma-1', 'mean-10-sigma-2'),
    assembler = c('metaspades', 'megahit'),
    error_profile = c('normal_err'),
    insert_size = c('mean350-stdev75')
)
train2b = do.call(expand.grid, params)
train2b

### Merged

train_all = rbindlist(list(train1, train2a, train2b), use.names=TRUE)
train_all

train$error_profile %>% table
train$insert_size %>% table
train$rep %>% table

train_all$error_profile %>% table
train_all$insert_size %>% table
train_all$rep %>% table

### Missing

nrow(train_all) - nrow(train)

train_f = train_all %>%
    anti_join.(train, c('rep', 'richness', 'read_length', 'read_depth', 
                        'abundance_distribution', 'assembler', 'error_profile',
                        'insert_size'))
df.dims(30)
train_f %>% arrange.(insert_size, error_profile, abundance_distribution, read_depth, read_length, richness)
df.dims()

F = file.path('/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/ftp/n9k-train_missing/todo.tsv')
train_f %>% 
    arrange.(insert_size, error_profile, abundance_distribution, read_depth, read_length, richness) %>%
    write_table(F)

# ResMiCo-SM

```
(snakemake) rick: /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/bin/old_compile/ResMiCo/ResMiCo-SM
$ screen -L -S rmc-train-c1 ./snakemake_sge.sh /ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/ftp/n9k-train_missing/1/config.yaml 30 -F

# ...and similar for other iterations
```

## Feature tables

base_dir = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/ftp/n9k-train_missing/'

keys = c('rep', 'richness', 'read_length', 'read_depth', 
         'abundance_distribution', 'assembler', 'error_profile',
         'insert_size')

### mean270-stdev50

# listing feature files
files = list_files(file.path(base_dir, 'mean270-stdev50'), 'feature_files.tsv')
files %>% length
files %>% head(n=3)

# reading, merging, & formatting
fillers1 = files %>%
    lapply(Fread) %>%
    rbindlist(use.names=TRUE) %>%
    mutate.(error_profile = 'cami_err',
            insert_size = 'mean270-stdev50')
fillers1

### mean350-stdev75

# listing feature files
files = list_files(file.path(base_dir, 'mean350-stdev75'), 'feature_files.tsv')
files %>% length
files %>% head(n=3)

# reading, merging, & formatting
fillers2 = files %>%
    lapply(Fread) %>%
    rbindlist(use.names=TRUE) %>%
    mutate.(error_profile = 'cami_err',
            insert_size = 'mean350-stdev75')
fillers2

### Overlap

fillers = rbindlist(list(fillers1, fillers2))
fillers

# checking overlap with to-do parameter sets
fillers %>%
    inner_join.(train_f, keys)

# any remaining?
train_f %>%
    anti_join.(fillers, keys)

### Copying samples

F = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/ftp/n9k-train_missing/todo.tsv'
todo = Fread(F)
todo

filler_dir = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/ftp/n9k-train_missing'
target_dir = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/ftp/n9k-train/features'

cp_fillers = function(x, filler_dir, target_dir){
    # files to copy
    file_names = c('features_binary', 'features.tsv.gz', 'stats', 'toc')
    # source
    src_dir = file.path(filler_dir, x['insert_size'], x['config'],
                        'features', x['richness'], x['abundance_distribution'],
                        x['rep'], x['read_length'], as.integer(x['read_depth']),
                        x['assembler']) 
    src_dir = gsub(' +', '', src_dir)
    if(! dir.exists(src_dir)){
        stop('Source directory does not exist: ', src_dir)
    }
    # destination
    dst_dir = file.path(target_dir, x['error_profile'], x['richness'],
                        x['abundance_distribution'], x['insert_size'],
                        x['rep'], x['read_length'], as.integer(x['read_depth']),
                        x['assembler'])
    # creating destination directory
    if(!dir.exists(dst_dir)){
        make_dir(dst_dir, quiet=TRUE)
    }
    # for each file, copy
    for(F in file_names){
        # source
        src_file = file.path(src_dir, F)
        if(! file.exists(src_file)){
            stop('Source file not found: ', src_file)
        }
        # destination
        dst_file = file.path(dst_dir, F)
        if(file.exists(dst_file)){
            stop('Dest. file already exists: ', dst_file)
        }
        # copying
        file.copy(src_file, dst_file)
    }
    return(dst_dir)
}
# running
dst_dirs = todo %>%
    apply(1, cp_fillers, filler_dir=filler_dir, target_dir=target_dir)
dst_dirs %>% length

### Creating new feature table

outfile = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/ftp/n9k-train'
outfile = file.path(outfile, 'feature_files.tsv')

train_all %>%
    select.(richness, abundance_distribution, insert_size, error_profile, 
            rep, read_length, read_depth, assembler) %>%
    mutate.(read_depth = as.integer(read_depth)) %>%
    mutate.(feature_file = file.path('features', error_profile, richness, 
                                     abundance_distribution, insert_size,
                                     rep, read_length, read_depth,
                                     assembler, 'features.tsv.gz')) %>%
    write_table(outfile)

# --OLD--

### TODO samples

F = '/ebio/abt3_projects/databases_no-backup/DeepMAsED/version2/resmico-sm/ftp/n9k-train_missing/todo.tsv'
todo = Fread(F)
todo

todo %>%
    semi_join.(fillers, keys) %>%
    arrange.(insert_size, error_profile, abundance_distribution, read_depth, read_length, richness)

# still to do
df.dims(30)
todo %>%
    anti_join.(fillers, keys) %>%
    arrange.(insert_size, error_profile, abundance_distribution, read_depth, read_length, richness)
df.dims()

todo %>%
    filter.(config == 5)

# sessionInfo

sessionInfo()
